cart section is divided into files
